snmp-trap-to-mqtt/requirements.txt
paho-mqtt==1.6.1
aiosnmp==0.7.2